+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+### v3.0.3 (unreleased)
+
+* Rename `--watch-and-serve` to `--server`
+* Update `MarkdownRenderer` defaults for `descriptive_class_title` and
+  `descriptive_class_title`
+* Added `MkdocsRenderer.content_directory` (defaults to `content`, used to be
+  `docs` before)
+* Renamed `MkdocsRenderer.clean_docs_directory_on_render` to `.clean_render`
+  and make it default to `False`
+* Fixed `--server` option, now reloads the `PydocMarkdown` config, does no
+  open the browser again and again
+* Changed: `MkdocsRenderer.mkdocs_config` can be set to `null`, the renderer
+  will then refrain from writing a `mkdocs.yml` configuration file into the
+  output directory
+
+### v3.0.2 (2020-05-16)
+
+* Fix `NameError` in `MkdocsRenderer`
+
 ### v3.0.1 (2020-05-16)
 
 * Added `--version` option to `pydoc-markdown` command.